return the key of the last row hashed

+++ src/row_printer.h
@@ -2,7 +2,7 @@
 template<class Row>
 struct RowPrinter {
 	void operator()(Row &row) {
-		for (int column_number = 0; column_number < row.n_columns(); column_number++) {
+		for (size_t column_number = 0; column_number < row.n_columns(); column_number++) {
 			cout << row.string_at(column_number);
 			if (column_number != row.n_columns() - 1) cout << "\t";
 		}
